requirements.txt
apipkg==1.5
appdirs==1.4.4
apt-clone==0.2.1
apturl==0.5.2
argcomplete==1.8.1
attrs==20.3.0
autopep8==1.5.4
beautifulsoup4==4.6.0
blurhash-python==1.0.2
Brlapi==0.6.6
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
command-not-found==0.3
configobj==5.0.6
cupshelpers==1.0
decorator==4.4.2
defer==1.0.6
distlib==0.3.1
execnet==1.7.1
fastapi==0.62.0
fastapi-utils==0.2.1
filelock==3.0.12
h11==0.11.0
h2==4.0.0
hpack==4.0.0
httplib2==0.9.2
Hypercorn==0.11.1
hyperframe==6.0.0
idna==2.6
iniconfig==1.1.1
louis==3.5.0
Mako==1.0.7
MarkupSafe==1.0
netifaces==0.10.4
onboard==1.4.1
packaging==20.8
pep8==1.7.1
pexpect==4.2.1
Pillow==8.0.1
pipenv==2020.11.15
pluggy==0.13.1
priority==1.3.0
progressbar==2.3
psutil==5.4.2
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.10.0
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pydantic==1.7.3
pygobject==3.26.1
PyICU==1.9.8
pyparsing==2.4.7
pytest==6.2.0
pytest-cache==1.0
pytest-pep8==1.0.6
python-apt==1.6.5+ubuntu0.2
python-debian==0.1.32
python-dotenv==0.15.0
python-gnupg==0.4.1
python-xapp==1.4.0
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
screen-resolution-extra==0.0.0
sessioninstaller==0.0.0
setproctitle==1.1.10
six==1.15.0
SQLAlchemy==1.3.20
starlette==0.13.6
system-service==0.3
systemd-python==234
toml==0.10.2
ubuntu-drivers-common==0.0.0
Ubuntu-Make==18.5
ufw==0.36
urllib3==1.26.2
uvicorn==0.13.0
validators==0.18.1
virtualenv==20.2.2
virtualenv-clone==0.5.4
wsproto==1.0.0
xkit==0.0.0
